@@ -1,16 +1,20 @@
 // auth.guard.ts
-import { Injectable } from '@angular/core';
-import { CanActivate, Router } from '@angular/router';
+import { Injectable, inject } from '@angular/core';
+import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
 import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthGuard implements CanActivate {
-  constructor(private authService: AuthService, private router: Router) {}
+  private authService = inject(AuthService);
+  private router = inject(Router);
 
-  canActivate(): boolean {
-    if (this.authService.getRole()) {
+
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean{
+    const urlSegments = state.url.split('/').filter(segment => segment);
+    const lastPath = urlSegments[urlSegments.length - 1];
+    if (this.authService.getRole() === lastPath) {
       return true;
     } else {
       this.router.navigate(['/login']);
